fix: use pipe2 to signal tracee to proceed to exec
Previously we raise SIGSTOP in tracee and then waitpid in tracer before seizing it, which yields a group stop(I think this is actually an UNDOCUMENTED behavior) after seizing. It hits some ptrace quirks.

For example, some child processes forked off after exec might caught this phantom ptrace group stop and hangs.

So this patch instead uses a pipe between tracer and root tracee to communicate when to proceed to exec.