fix(bpf): move fd_event usage up to save work for verifier
In compiled assembly, fd_event argument of read_send_path
is spilled onto stack. Since it is a pointer to BPF map element,
this requires the verifier to track it on the stack across the long,
complex body of read_send_path(which is almost an implementation
of bpf_d_path helper in bpf) that contains several function calls.

Since it does not depend on most of the earilier code, move it up
so that the verifier could do less work here.

Even though the register is spilled when compiled with both clang
20 and 21, before or after the bisected clang commit.
This still workarounds #105.
I am not sure if there's a problem with clang as the assembly diff
is too painful to read.